Conflicts:
	scilab/CHANGES_5.4.X
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln
	scilab/configure
	scilab/modules/call_scilab/src/c/call_scilab.c
	scilab/modules/gui/src/java/org/scilab/modules/gui/bridge/label/SwingScilabLabel.java
	scilab/modules/ui_data/src/java/org/scilab/modules/ui_data/filebrowser/ScilabFileBrowserHistory.java
	scilab/modules/ui_data/src/java/org/scilab/modules/ui_data/filebrowser/SwingScilabTreeTable.java
	scilab/modules/xcos/src/java/org/scilab/modules/xcos/XcosTab.java

+++ scilab/modules/fileio/src/c/sendtorecyclebin.c
@@ -34,7 +34,7 @@ BOOL sendToRecycleBinW(const wchar_t* wfilename)
     if (wfilename)
     {
         SHFILEOPSTRUCTW opFile;
-        int len = wcslen(wfilename) + 2;
+        size_t len = wcslen(wfilename) + 2;
         /* we add 00 at the end of filename */
         /* required by SHFileOperationW */
         wchar_t* csrc = (wchar_t*)CALLOC(len, sizeof(wchar_t));
